requirements.txt
# web 服务框架
fastapi
uvicorn
uvloop

# Telegram userbot
pyrogram
tgcrypto

# 数据库
sqlalchemy[asyncio]
asyncpg

# 配置文件
pydantic-settings
pydantic[dotenv]
python-dotenv

# 判断文件类型
python-magic

# 日志
loguru
richuru